unify original_prompt output across attacks
The `original_prompt` field of `SingleAttackRunResult` should contain the full conversation from the dataset used to run an attack (including the affirmative target, if available). We want this because reproducing results will be easier if the optimization target that was used is included  in the attack result artifact. It matters less for attack which don't rely on dataset-provided targets like e.g. actor attack but it'd still be nice to have it be consistent